// src/app/shared/models/layout.model.ts
export enum LayoutType {
  SINGLE = 'single',
  HORIZONTAL_TWO = 'horizontal-two',
  VERTICAL_TWO = 'vertical-two',
  ONE_TOP_TWO_BOTTOM = 'one-top-two-bottom',
  TWO_TOP_ONE_BOTTOM = 'two-top-one-bottom',
  ONE_LEFT_TWO_RIGHT = 'one-left-two-right',
  TWO_LEFT_ONE_RIGHT = 'two-left-one-right',
  FOUR = 'four'
}